Refactored query forms
